"Update communication and control functions to use TCP; remove obsolete functions"
This commit updates the `communication` and `control` packages to use TCP for communication with the CLP. The `Tcp` function in `communication` package is updated to connect to the CLP via TCP and send/receive data. The `TcpVerification` function in `control` package is removed as it is no longer needed. A new function `Authentication` is added to `control` package for user authentication. The `VerifyUser` function is also added to `control` package for verifying the user's credentials. The `TcpConnection` function is added to `communication` package for establishing a TCP connection with the CLP. The `CheckTcpCommunication` function is removed as it is no longer needed. The `tagDB` package is still being imported but not used in this commit.